pyproject.toml
[tool.poetry]
name = "classroom-lm-poc"
version = "0.1.0"
description = ""
authors = ["Bharath Sriraam R R <albharath.1305@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.3.1"
numpy = "^1.26.4"
langchain = "^0.2.3"
langchain-openai = "^0.1.8"
instructor = "^1.3.2"
ipykernel = "^6.29.4"
python-dotenv = "^1.0.1"
langgraph = "^0.0.65"
chromadb = "^0.5.0"
openai = "^1.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
